10 Coordinates: int -> float

+++ app/crawiot_common.h
@@ -1,6 +1,6 @@
 #ifndef APP_CRAWIOT_COMMON_H
 #define APP_CRAWIOT_COMMON_H
 
-int calculate_diff(int current, int target);
+float calculate_diff(float current, float target);
 
 #endif //APP_CRAWIOT_COMMON_H
